refactor openstack_manager to openstack_cloud module
