+++ src/charm.py
@@ -35,7 +35,6 @@ from ops.main import main
 from ops.model import ActiveStatus, BlockedStatus, MaintenanceStatus
 
 import metrics
-import openstack_manager
 from charm_state import (
     DEBUG_SSH_INTEGRATION_NAME,
     CharmConfigInvalidError,
@@ -60,6 +59,7 @@ from event_timer import EventTimer, TimerStatusError
 from firewall import Firewall, FirewallEntry
 from github_client import GithubClient
 from github_type import GitHubRunnerStatus
+from openstack_cloud import openstack_manager
 from runner import LXD_PROFILE_YAML
 from runner_manager import RunnerManager, RunnerManagerConfig
 from runner_manager_type import FlushMode
