+++ src/charm_state.py
@@ -17,6 +17,7 @@ from ops import CharmBase
 from pydantic import AnyHttpUrl, BaseModel, Field, ValidationError, root_validator
 from pydantic.networks import IPvAnyAddress
 
+import openstack_cloud
 from errors import OpenStackInvalidConfigError
 from firewall import FirewallEntry
 from utilities import get_env_var
@@ -24,10 +25,11 @@ from utilities import get_env_var
 logger = logging.getLogger(__name__)
 
 ARCHITECTURES_ARM64 = {"aarch64", "arm64"}
-
 ARCHITECTURES_X86 = {"x86_64"}
 
 CHARM_STATE_PATH = Path("charm_state.json")
+
+OPENSTACK_CLOUDS_YAML_CONFIG_NAME = "experimental-openstack-clouds-yaml"
 
 
 StorageSize = str
@@ -162,54 +164,6 @@ def _valid_storage_size_str(size: str) -> bool:
     return size[-3:] in valid_suffixes and size[:-3].isdigit()
 
 
-OPENSTACK_CLOUDS_YAML_CONFIG_NAME = "experimental-openstack-clouds-yaml"
-CLOUDS_YAML_PATH = Path(Path.home() / ".config/openstack/clouds.yaml")
-
-
-def _validate_openstack_cloud_config(cloud_config: dict) -> None:
-    """Validate the format of the cloud configuration.
-
-    Args:
-        cloud_config: The configuration in clouds.yaml format to validate.
-
-    Raises:
-        InvalidConfigError: if the format of the config is invalid.
-    """
-    # dict of format: {clouds: <cloud-name>: <cloud-config>}
-    try:
-        clouds = list(cloud_config["clouds"].keys())
-    except KeyError as exc:
-        raise OpenStackInvalidConfigError("Invalid clouds.yaml.") from exc
-
-    if not clouds:
-        raise OpenStackInvalidConfigError("No clouds defined in clouds.yaml.")
-
-
-def _write_openstack_config_to_disk(cloud_config: dict) -> None:
-    """Write the cloud configuration to disk.
-
-    Args:
-        cloud_config: The configuration in clouds.yaml format to write to disk.
-    """
-    CLOUDS_YAML_PATH.parent.mkdir(parents=True, exist_ok=True)
-    CLOUDS_YAML_PATH.write_text(encoding="utf-8", data=yaml.dump(cloud_config))
-
-
-def initialize_openstack(cloud_config: dict) -> None:
-    """Initialize Openstack integration.
-
-    Validates config and writes it to disk.
-
-    Args:
-        cloud_config: The configuration in clouds.yaml format to apply.
-
-    Raises:
-        InvalidConfigError: if the format of the config is invalid.
-    """
-    _validate_openstack_cloud_config(cloud_config)
-    _write_openstack_config_to_disk(cloud_config)
-
-
 class CharmConfig(BaseModel):
     """General charm configuration.
 
@@ -290,21 +244,21 @@ class CharmConfig(BaseModel):
         denylist = cls._parse_denylist(charm)
         dockerhub_mirror = cls._parse_dockerhub_mirror(charm)
 
-        openstack_clouds_yaml_str = charm.config.get(OPENSTACK_CLOUDS_YAML_CONFIG_NAME)
+        openstack_clouds_yaml_str = charm.config.get("experimental-openstack-clouds-yaml")
         if openstack_clouds_yaml_str:
             try:
                 openstack_clouds_yaml = yaml.safe_load(openstack_clouds_yaml_str)
             except yaml.YAMLError as exc:
-                logger.error("Invalid openstack-clouds-yaml config: %s.", exc)
+                logger.error("Invalid experimental-openstack-clouds-yaml config: %s.", exc)
                 raise CharmConfigInvalidError(
-                    "Invalid openstack-clouds-yaml config. Invalid yaml."
+                    "Invalid experimental-openstack-clouds-yaml config. Invalid yaml."
                 ) from exc
             if (config_type := type(openstack_clouds_yaml)) is not dict:
                 raise CharmConfigInvalidError(
                     f"Invalid openstack config format, expected dict, got {config_type}"
                 )
             try:
-                initialize_openstack(openstack_clouds_yaml)
+                openstack_cloud.initialize(openstack_clouds_yaml)
             except OpenStackInvalidConfigError as exc:
                 logger.error("Invalid openstack config, %s.", exc)
                 raise CharmConfigInvalidError(
