+++ src/openstack_cloud/openstack_manager.py
